Tasks now works on iOS
The way that the LI was taken from the path of the event did not work
on mobile. Instead, the list item is now given in an arrow function of
a click handler on the checkbox.

This also means there is no longer the global click event listener.
Overall, some due clean-up happened.

Fixes #22