Tasks are now identified by - [ ] #task
The old system with `- TODO` and `- DONE` confused people. The
markdown-default checklist `- [ ]` is what people expect, so it shall
be what we use.
Furthermore, there will be a metadata-cache update to obsidian, where I
will be able to query for checklist items.

Still, I don't want *all* checklist items to be tracked tasks.
Therefore, tasks need to have a `#task` tag, e.g. `- [ ] #task`, to be
identified as a task.

The tag is a config setting, but it is not yet exposed in the settings
menu.

The two statuses are currently "todo" `[ ]` and "done" `[x]`.

Using a tag to identify tasks should also make it easier to find them
with the cache provider from obsidian.

Fixes #19