Transclusion unloads properly
The transclusion event listener is now removed properly when the
element is no longer rendered and added again when it is rendered anew.

The issue was that I did not send the markdown render child's private
`containerEl` property. If that property is `null`, the component gets
unloaded.

The change fixes #14.

Furthermore, from testing throughout the day, it seems that this commit
together with the earlier `cachedRead` commit 4a2f3b8 fixes #15.