Initial version with API 12
Obsidian 12 comes with a new API for plugins. This is a complete
re-write based on the new API.

This version is on feature-parity with the previous version.

However, I still want to add global filters to enable the option to
filter out unwanted items from inline tasks and query blocks.