Obsidian 12 rewrite (#34)
* Initial version with API 12

Obsidian 12 comes with a new API for plugins. This is a complete
re-write based on the new API.

This version is on feature-parity with the previous version.

However, I still want to add global filters to enable the option to
filter out unwanted items from inline tasks and query blocks.

* Tasks can now be filtered globally

Added a new setting to globally filter tasks by a string.
This can be used, for example, to filter only checklist items that have
a certain tag, e.g. `#task`. Doesn't have to be a tag, though.

* Update readme for obsidian 12 rewrite