Nested global filters no work as intended
Previously, this bug existed:
Given that I use a global filter `#task`
When I nest tasks under a general checklist item
Then the markdown preview of the file is messed up
But instead it should render as expected

The rendering in the markdown preview adds the text of the first task
in the sub-list to the checklist item (replacing the checklist item's
text). All other checklist items/tasks "move up by one" content-wise.

This doesn't work:
```
- [ ] A list of tasks:
    - [ ] #task This is my first task
    - [ ] #task This is my second task
```
This works:
```
- [ ] #task A list of tasks:
    - [ ] #task This is my first task
    - [ ] #task This is my second task
```

Fixes #42