Cache now loads tasks correctly after load
The cache wasn't re-loaded immediately if the plugin was activated
after the initial loading of obsidian, because the metadata cache's
"resolve" wouldn't fire until a file was edited.

Now, the cache loads twice in that case. Once on instantiation and once
again after the first metadata cache's resolve. This is necessary to
still re-load the task cache after the first metadata cache's resolve
when starting obsidian, where the initial task cache's load would load
0 tasks from the cold metadata cache.

Fixes #152