Recurrence now works as intended
Subsequent due dates of recurring tasks weren't always on the expected
date.

The reason was that the `RRule` was constructed initially without
setting `dtStart`. As a result, properties like `bymonthday` were set
to "today" instead of the correct day based on the due date.

The fix is to parse an `options` object first and set `dtStart` on the
`options` object *before* constructing the `RRule`.

All code related to `RRule` moved into the new `Recurrence` class for
consistency.

Closes #261, closes #324, closes #352