Spaced tasks are now rendered correctly in preview
When checklists with tasks contain an empty line in between list items,
the list is rendered with spacing around each list item. This spacing
resulted in the Tasks plugin not picking up the correct content of each
LI element.

The plugin now correctly searches for the task content in the LI
element.

Note that the tasks won't be rendered with spacing around them like
"normal" list items.

Fixes #313