docs: Fix occasional incorrect images in user docs & plugin store.
The User Guide and README.md were showing images from the main
branch, whereas the docs are published from the gh-pages branch.

This meant that if existing screenshots were updated on main, they would
be shown in the documentation and community plugin store, prior to the
new feature being released to users.

For consistency, I've opted for making all images shown in this repo
from the gh-pages repo, including the README. This fixes the related
problem that the Obsidian Community Plugin store was showing an image
of a feature that has not yet been released.