Ensure monthly and yearly recurrences
When a recurrence rule states "every month" and the next month does not
have the day of the month that the task is scheduled, that month is
skipped and the next valid month is used instead.

For example: When I mark a task that recurs "every month" done on
January 31, the next occurrence will be scheduled on March 31, since
February does not have 31 days. This is in line with the rrule
specification, but not what users expect. This can be triggered by
accident, for example when the task is based on the done date and the
done date is on the 31st.

@claremacrae architected a fix that I implemented with this change.

To fix the issue, the number of expected skipped months is parsed
(defaults to one).
While the actualy number of skipped months is higher, the base date is
adjusted one day into the past. In practice, this means that the example
from above would no schedule the recurring task on the 28th of
February (or the 29th for a leap year).

The same applies for yearly recurrences, for example based on the 29th
of February.

This is a proof of concept and is still missing:

1. Documentation in code
2. End user documentation to explain the behavior
3. Extensive testing to ensure this change does not introduce regressions

Fixes #614, #1087