Ensure monthly and yearly recurrences (#1197)
* Ensure monthly and yearly recurrences

When a recurrence rule states "every month" and the next month does not
have the day of the month that the task is scheduled, that month is
skipped and the next valid month is used instead.

For example: When I mark a task that recurs "every month" done on
January 31, the next occurrence will be scheduled on March 31, since
February does not have 31 days. This is in line with the rrule
specification, but not what users expect. This can be triggered by
accident, for example when the task is based on the done date and the
done date is on the 31st.

@claremacrae architected a fix that I implemented with this change.

To fix the issue, the number of expected skipped months is parsed
(defaults to one).
While the actualy number of skipped months is higher, the base date is
adjusted one day into the past. In practice, this means that the example
from above would no schedule the recurring task on the 28th of
February (or the 29th for a leap year).

The same applies for yearly recurrences, for example based on the 29th
of February.

This is a proof of concept and is still missing:

1. Documentation in code
2. End user documentation to explain the behavior
3. Extensive testing to ensure this change does not introduce regressions

Fixes #614, #1087

* Adding code comments

* Adding tests from @claremacrae

* Fixed edge cases across years

See tests for failing edge cases.

* Document recurrence edge cases

* test: Test recurrence corrects next date with 'when done'

This test the interaction of 'when done' and the fix for the bug
that months or years could be skipped when the next reference
date was invalid.

* vault: Add Recurrence handling invalid dates.md

This was useful for exploratory testing of recurrence,
for #1087: 'Recurrence with 6 months recurs every year'

* vault: Capture current behaviour of 'very month on the'

When next iteration lands on a non-existent date.

* vault: Simplify re-testing of recurrence on fixed days

* vault: Demonstrate a bug that 'every month on the 31st' gets stuck

With new recurrence code, if 'every month on the 31st' lands on an
invalid date, all future new tasks are stuck on the same start date.

* vault: Remove misleading task descriptions

* vault: Enable outline core plugin

* vault: Set up manual testing of 'every month on the last'

* vault: Capture current behaviour of 'every month on the last'

* vault: Capture new behaviour of 'every month on the last'

It's good, and unchanged.

* test: Guard against recurrence tests with no expectations set

* docs: Document behavior if reference date is invalid

* docs: Remove obsolete instruction about ordering of recurrence rule.

* fix: Prevent 'every month on the 31st' sticking on 31st January

* vault: Update sample file with fixed behavior for 'every month on the 31st'

* docs: Add major new section: 'How the New Date is Calculated: Repeating Monthly'

* docs: Add new section: 'Technical Details'

* docs: Add markdown codeblocks to tasks in new sections

* docs: Use more representative done does in sample tasks

* docs: Warn against use of 'every month on the 31st'

* docs: Warn near top of file about skipped recurrences

* test: Add test to see whether 'every year on' needs special case. It seems not.

* refactor: code review: Remove repeated calls to Recurrence.toText()

Co-authored-by: Clare Macrae <github@cfmacrae.fastmail.co.uk>