fix: Prevent unreported bug that 'priority is abovemedium' worked
fix: Prevent unreported bug that 'priority is abovemedium' was valid

It was interpreted as 'priority is above medium', but it was confusing
that it was valid.

This was reported by @Cito in the code review for #1365,
along with a proposed general fix.

In this step I make the minimal change needed to just fix this one issue.

Co-Authored-By: Christoph Zwerschke <cito@online.de>