refactor: Add StatusConfiguration & Status from obsidian-tasks-x
Behaviour is unchanged.

Credit: This is a copy of the code and tests in
@sytone's fork https://github.com/sytone/obsidian-tasks-x
on the main-tasks-sql branch.

Note: Initially only Status.DONE and Status.TODO have been copied over,
to maintain parity with the existing behaviour in Tasks and keep this
as a pure refactoring.

The revision copied from:
https://github.com/sytone/obsidian-tasks-x/commit/2c0b659457cc80806ff18585c955496c76861b87

Co-Authored-By: Sytone <1399443+sytone@users.noreply.github.com>