refactor: Add most of StatusRegistry from obsidian-tasks-x
This copies just the parts of StatusRegistry that I believe
are necessary to be able to incrementally refactor away
Task.originalStatusCharacter and replace it with Task.status.indicator.

(My initial attempts to use Task.status.indicator failed because
the new Status implementation did not easily support the use
of unrecognised status characters. I then found myself starting to write
interim code that looked suspiciously like what StatusRegistry
did already, so I will incrementally adjust the existing code
to start using StatusRegistry, even though only two status types
are currently supported. This will then make the later adoption
of StatusRegistry easier, I believe.)

The behaviour of this project is unchanged.

Credit: This is a partial copy of the code and tests in
@sytone's fork https://github.com/sytone/obsidian-tasks-x
on the main-tasks-sql branch.

Note: This adds Status.EMPTY.

The revision copied from:
https://github.com/sytone/obsidian-tasks-x/commit/2c0b659457cc80806ff18585c955496c76861b87

Co-Authored-By: Sytone <1399443+sytone@users.noreply.github.com>