fix: Fixed confusing nested Status objects returned from StatusRegistry.
I had found tests were failing confusingly because Status values
returned from StatusRegistry had a duplicate nested configuration
field, and I didn't understand why.

It turned out that the code in obsidian-tasks-x was offering the convenience
of adding either StatusConfiguration or Status to the StatusRegistry,
but was not checking the type passed in. This meant that if a Status
was added, what was saved was Status(Status(StatusConfiguration)).

This is why the code from obsidian-tasks-x which I added in
284e00c13efb1b8ec40e365eaef18a2524e74080 had to do all the
converting from:
    expect(task!.status).toStrictEqual(Status.DONE);
to:
    expect(task!.status.indicator).toStrictEqual(Status.DONE.indicator);