refactor: Adopt obsidian-tasks-x settings mechanism
In this first revision, the current Tasks settings code is
unchanged, but new placeholder text is added, for where the
new status settings will go.

The text for that new placeholder is in settingsConfiguration.json.

Things to investigate in subsequent revisions:
- The style of the new descriptive text is a bit bolder than
  in obsidian-tasks-x.
- There is a temporary unwanted side-effect that existing settings
  are now duplicated in a 'generalSettings' section. I'll remove that
  in a separate commit.
- Reworking all the pre-existing settings to move them over to
  settingsConfiguration.json.

Credit: This is a derived copy of the code in
@sytone's fork https://github.com/sytone/obsidian-tasks-x
from its main-tasks-sql branch.

The direct revision copied from is in @claremacrae's fork of sytone's obsidian-tasks-x,
with some refactoring to extract methods for re-use and to re reduce repetition,
on the branch 'refactor-settings-code-for-reuse':
https://github.com/claremacrae/obsidian-tasks-x/commit/4f59b9dd16b4ede25ecc9957deb51967f99a1106

The original revision copied from is:
https://github.com/sytone/obsidian-tasks-x/commit/2c0b659457cc80806ff18585c955496c76861b87

Co-Authored-By: Sytone <1399443+sytone@users.noreply.github.com>