fix: Prevent duplicate values in user settings
When I start porting the pre-existing settings sections
to the new settingsConfiguration.json file, I will
have to deal with its duplication of pre-existing
settings values. But for now, we don't need to do that,
and I don't want to clutter up users settings yet,
in case I can find an alternative way of dealing
with this.