refactor!: Add StatusSettings interface & rename statuses setting
Note: Any previously saved custom statuses will need to be re-created,
as the settings format has changed.

This is a step towards pulling all the code for handling status settings
in to one self-contained, testable location.

StatusSettings is now an interface. Doing this broken the reading
of previously-saved custom status settings. Which is fine, as this is
an un-released feature.