fix: Prevent Tasks getting stuck 'Loading Tasks...'
Also add a message to users if other file-reading issues occur in future.

Obsidian CachedMetadata was telling us that there is a task on lineNumber,
but there are not that many lines in the file.

This was the underlying cause of all the 'Stuck on "Loading Tasks..."' messages,
as it resulted in the line 'undefined' being parsed.

Somehow the file had been shortened whilst Obsidian was closed, meaning that
when Obsidian started up, it got the new file content, but still had the old cached
data about locations of list items in the file.

Fixes #829
Fixes #1033
Fixes #1048
Fixes #1072