Implement guard on task write to file
Currently, the unique identifier (UID) for tasks
are a source file, a section, and an index into
the array of global-filter-matching tasks within
that section.

This diff does not change how tasks are uniquely
identified.

However, it does implement a guard when overwriting a
task in some file to ensure the line that is being
overwritten has similar content to the line that
was used to create the Task in the first place.

The world "similar" is used since the line is
permitted to have different content, so long as that
line represents a Task with identical data as per
the definition of Task.identicalTo