chore: Set up mock mechanism for the toggling code
This is part of a desperate bid to create some tests
for the many failure cases of replaceTaskWithTasks().

It hard-codes in to tryRepetitive() that if the correct
line cannot be found, then to write out a JSON representation
of the data being used to locate the line being toggled.

This way, tests can be written using the saved mock data.