fix: Prevent exception if toggling in unsaved file in Reading mode
See the comment https://github.com/obsidian-tasks-group/obsidian-tasks/issues/1680#issuecomment-1457826313
in #1680.

If lines had been deleted from a file, and it was not yet saved when user
then toggled a line near the end of the file in Reading mode,
there could be an access of a non-existent line number.

The resultant output was:

plugin:obsidian-tasks-plugin:17605 Uncaught (in promise) TypeError: Cannot read properties of undefined (reading 'includes')
    at eval (plugin:obsidian-tasks-plugin:17605:14)
    at Generator.next (<anonymous>)
    at fulfilled (plugin:obsidian-tasks-plugin:173:24)