feat: Display a meaningful Help message if Task cannot be found.
Also increase the message duration to 15 seconds, to allow user time
to read this longer message.

Example output:

Tasks: Could not find the correct task line to update.

The task line not updated is:
- [ ] #task task2b ^ca47c7

In this markdown file:
"Manual Testing/Task Toggling Scenarios/Embed Task in to Note.md"

Note: further clicks on this checkbox will usually now be ignored until the file is opened (or certain, specific edits are made - it's complicated).

Recommendations:

1. Close all panes that have the above file open, and then re-open the file.

2. Check for exactly identical copies of the task line, in this file, and see if you can make them different.