docs: Rename documentation files to match title
The file names will show up in the site index when published,
and the names with spaces are a lot more readable.

One exception: 'Aura Theme (Dark mode only)' becomes 'Aura Theme',
because I know that it will soon be updated to reflect that
it now supports Light mode too.