feat: add sql query engine
Simplified implementation of the SQL based query engine, I have been
using this for a while now and not had any major issues outside of
grouping. This is the most minimal level of changes to have it
integrated into the Tasks plugin.

Grouping will be a work in progress as the Tasks model of grouping and
rendering of groups is static and based off well known columns. The SQL
based approach is a dynamic tree of objects. Simple single level
grouping work for the majority of areas.

The TaskRecord is the simpler representation of the task that the
queries can run against, it flattens out some of the properties for
ease and can be expanded as needed for other properies. Basic functions
are implemented. Integration with the CustomJS plugin is commented out
and can be added later. This is the same case as for the use of page
properties in the query as the query engine needs access to the front
matter of the page it is running on, that way page level properties can
be used in the queries simplifying the templating process.