docs: cleaning up docs for date found some other issues
cleaned up the docs for the move from Date to Moment
and hit a few other issues so refactored to break up the
logic. Refactored Layout options as it was duplicated
between both engines. Added in the logic to handle custom
functions as CustomJS integration needs future review. This
needed the addition of the multiline directive to support
that and enable all my queries to run and a bit of a move.

While debugging saw that the logging options were not
in the settings so added them back and fixed some naming.
Updated the tests to validate refactors and correct multiline
parsing.

Added more comments to make the code more readable.