fix: Correct nested headings when multiple groups are used.
In 809aa9d4645b0799621048b53ef42fb7f2a3f02f I moved the sorting
of groups to a different location, and didn't spot the change
in behaviour, because I was now sorting the groups after the display headings
had been calculated, meaning that in some cases the wrong headings were displayed.

Now, the groups are sorted before calculating which headings to display.