test: Fix fromLine() test function behaviour when no heading
The actual behaviour of the Tasks plugin when there is no available
preceding heading is to use null, not an empty string.

This makes the test helper fromLine() in TestHelpers.ts behave the same way.

All tests still pass.

(TestBuilder already used null as the default heading, so no extra work was
needed there.)