refactor: Add the 'group by' field name to GroupDisplayHeading
See #1867. This should enable styling to be done, depending on the type of group in each group heading.

I called it 'property' for consistency with the same information elsewhere
in the grouping code, even though this is a hangover from an earlier implementation,
and 'fieldName' is probably a more accurate name.