fix: Enable 'High priority' to be selected easily in Auto Suggest (#1996)
Fixes #1943

The order of priorities in auto suggestion is now

- high
- medium
- low
- highest
- lowest

This retains the original order, for people's muscle memory. And it means that the
longer names are after the shorter ones, which also makes it possible to type both,
and it means that typing parts of names selects the same result as before
highest and lowest were added.