Toggle task done command adds global filter text
Fixes 1320

I like to use the "Tasks: Toggle task done" command to create tasks, but
since I have a global filter set I had to manually type the global
filter each time. This PR changes the behavior so that if the user has a
global filter set then the global filter is inserted when the
"Tasks: Toggle task done" is run.