Add more tests
Also fix the addition of an empty space when the task doesn't have any
description besides the global filter. Otherwise some of test cases felt
broken because an extra space would be inserted.

'- |#task' would become '- [ ]  #task|' instead of '- [ ] #task|'

Also fix the case where the task description already includes the global
filter. In that case it is not necessary to add the global filter.