chore: Specify supported node versions in package.json (#2065)
* Specify supported node versions

Obsidian tasks does not support older versions of nodejs and the tests
do not pass with them. Specify the supported version of node in
package.json to ensure that contributors are running with an up to date
version of node.js

https://docs.npmjs.com/cli/v7/configuring-npm/package-json#engines

* chore: Prevent node 17 being used

It currently gives:

error expect@29.5.0: The engine "node" is incompatible with this module.
Expected version "^14.15.0 || ^16.10.0 || >=18.0.0". Got "17.9.1"

---------

Co-authored-by: Clare Macrae <github@cfmacrae.fastmail.co.uk>