test: - Refactor making of input data in StatusRegistryReport test
It now goes StatusCollection -> StatusSettings -> StatusRegistry

This is a step to making it easier to add more test cases.

The change in approved output shows that StatusRegistry
actually adds its 4 default statuses in a different order
to those in StatusSettings.

I viewed the two markdown diagrams - before and after - in
Obsidian and they were identical.