test: Add failing test to demo that changing status in test is broken
I have spent hours and hours trying different ways to get the
fireEvent() in tests to call to select a different option in the Status menu
in the Edit Task model.

Fundamentally, when the value is an object rather than a string,
our current test mechanism is not reflecting any changed selection.