feat: Allow user to control the date used to calculate next recurrence date.
When marking 'when done' recurring tasks as Done in the Edit Task modal,
the date of the next occurrence is now calculated from the Done date field.

The Done date value defaults to the current date, but can then be edited
before clicking Apply.

This might be useful if you realise that you had forgotten to mark
a recurring task as Done on the actual day that you completed it,
and so you would like the new instance to be created based on the day
you really completed the task, rather than today when you are marking
it as done.

(Today's date is still used for an Created day on a new recurrence, though.)