feat!: Disallow mixing () and "" delimiters in Boolean queries
This reduces the number of scenarios in which we will accidentally
treat a character in the filter as a delimiter.

For example, it fixes this filter:
    (description includes "hello world") OR (description includes "42")

Since the documentation discouraged use of " as a delimiter, it is
unlikely to break any/many real-world searches - but it is a breaking change.