fix: @ Rewrite isAFilter() to correct the processing of filters that are all capitals
This is tested, but I decided to use Arlo Belshee's @ symbol to signify the scale of the change.

This set of regular expressions was built up empirically through a lot of iteration,
over a very thorough set of sample Boolean filters, in order to detect all the outputs
from splitLine() that were not actually Tasks filters.

In this commit, I tried breaking it down in to smaller chunks, but I ended up feeling
that it was more useful to commit the corrected implementation, so that the changes to
tests in this commit usefully showed the small change in behaviour of the rewrite.