refactor: - Make handling of spaces after recurrence emoji consistent
In #900 I changed a load of ' ?' to ' *', to relax the reading of spaces.
I didn't do it for recurrenceRegex, presumably because extra spaces
were permitted by a later part of recurrenceRegex, so the tests did
not demand it.

However, I'm now converting the ? to a * for consistency with
the other regular expressions, to simplify later refactoring steps.