refactor: ! Replace displayName value in TaskFormat with function getDisplayName()
This will prevent the plugin crashing during starting when
TASK_FORMATS is initialised before translation is set up.

But it also has the advantage that if the language is changed during a program
session, the latest translation would be displayed.
(I've no idea if changing language in an Obsidian session is even possible, though...)