fix: Reinstate date picker on iPhone - week starts on Monday
This fixes #3341.

Note that the date picker in Reading Mode and Tasks query Results
remains hard-coded to start on Monday, because it turned out
that the code added in #3259 to address #3239 works well on Android,
but gave an error on iPhone (and probably iPad too).

Thanks to @esm7 for suggesting the fix.

Co-Authored-By: esm7 <esm7@users.noreply.github.com>