fix: Guard against parallel rendering in requestAnimationFrame()
Empirically, this prevents the incorrect rendering that is reported in #3715,
when triggered by using the the "Refresh Any View" button.

We are unsure whether this also fixes the same corrupt rendering that I had
observed often on my iPhone.

So I'm merging this now so that I can test the behaviour on iPhone for a
few days.

Co-Authored-By: Ilyas Landikov <93825870+ilandikov@users.noreply.github.com>